fixing errror + some things :)

+++ uniborg/_core.py
@@ -8,6 +8,7 @@ import os
 import traceback
 from datetime import datetime
 from uniborg import util
+thumb_image_path = Config.TMP_DOWNLOAD_DIRECTORY + "/thumb_image.jpg"
 
 logging.basicConfig(
     format='[%(levelname) 5s/%(asctime)s] %(name)s: %(message)s',
@@ -56,20 +57,26 @@ async def send_plug_in(event):
     message_id = event.message.id
     input_str = event.pattern_match["shortname"]
     the_plugin_file = "./stdplugins/{}.py".format(input_str)
-    start = datetime.now()
-    await event.client.send_file(
-        event.chat_id,
+    thumb = None
+    if os.path.exists(thumb_image_path):
+        thumb = thumb_image_path
+    if os.path.exists(the_plugin_file):
+    	start = datetime.now()
+    	await event.client.send_file(
+    	event.chat_id,
         the_plugin_file,
         force_document=True,
         allow_cache=False,
         caption="©️ @LazyAF_Pepe",
-        reply_to=message_id
-    )
-    end = datetime.now()
-    time_taken_in_ms = (end - start).seconds
-    await event.edit("Ok, BTC Uploaded {} in {} seconds".format(input_str, time_taken_in_ms))
-    await asyncio.sleep(DELETE_TIMEOUT)
-    await event.delete()
+        reply_to=message_id,
+        thumb=thumb)
+        end = datetime.now()
+        time_taken_in_ms = (end - start).seconds
+        await event.edit("Ok, BTC Uploaded {} in {} seconds".format(input_str, time_taken_in_ms))
+        await asyncio.sleep(DELETE_TIMEOUT)
+        await event.delete()
+    else:
+        await event.edit("**404:** `File Not Found`")    
 
 
 @borg.on(util.admin_cmd(pattern="install"))
